requirements.txt
aac-metrics==0.6.0
accelerate @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_accelerate_1747321561/work
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
arm-mango==1.5.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
botorch @ file:///home/conda/feedstock_root/build_artifacts/botorch_1746576285572/work
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
clip-score==0.2.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.14
decorator==5.2.1
Deprecated @ file:///opt/conda/conda-bld/deprecated_1659726428456/work
dill==0.3.8
docker-pycreds==0.4.0
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1739083394647/work
evaluate==0.4.5
exceptiongroup==1.2.2
executing==2.2.0
filelock @ file:///croot/filelock_1744281381737/work
flake8==7.3.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
gpytorch==1.14
hf-xet==1.1.5
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata @ file:///croot/importlib_metadata-suite_1732633488278/work
ipykernel==6.29.5
ipython==8.35.0
jaxtyping @ file:///home/conda/feedstock_root/build_artifacts/jaxtyping_1747196441045/work
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib @ file:///croot/joblib_1718217211762/work
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.14.3
linear_operator==0.6
llvmlite==0.44.0
lpips==0.1.4
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
msclap==1.3.4
msgpack==1.1.1
multidict==6.4.3
multipledispatch @ file:///home/conda/feedstock_root/build_artifacts/multipledispatch_1721907546485/work
multiprocess==0.70.16
mypy_extensions @ file:///home/conda/feedstock_root/build_artifacts/mypy_extensions_1745776566517/work
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numexpr @ file:///croot/numexpr_1730215937391/work
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
opentelemetry-api @ file:///croot/opentelemetry-api_1743171649300/work
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.1
pyparsing==3.2.3
pyre-extensions @ file:///home/conda/feedstock_root/build_artifacts/pyre-extensions_1734755991108/work
pyro-api @ file:///home/conda/feedstock_root/build_artifacts/pyro-api_1614940619870/work
pyro-ppl @ file:///home/conda/feedstock_root/build_artifacts/pyro-ppl_1734535654352/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pythonwrench==0.3.0
pytorch-fid-wrapper==0.0.4
pytorch-ignite==0.5.2
pytorch-sfid @ git+https://github.com/evenmn/pytorch-sfid@a5797123bc22d0652de0d7623c20bef52e6796bc
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1749487690589/work/dist/scikit_learn-1.7.0-cp310-cp310-linux_x86_64.whl#sha256=70f1a75eec730ad39d1de7603a85e2d091694943733e03fefa40748c9e8d7a1d
scipy @ file:///croot/scipy_1747238027288/work/dist/scipy-1.15.3-cp310-cp310-linux_x86_64.whl#sha256=4055d6b52bbbfc19594ac7c25008889945de175e153afc0935b6415716fa348e
sentence-transformers==5.0.0
sentry-sdk==2.29.1
setproctitle==1.3.6
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
sympy @ file:///croot/sympy_1738108488918/work
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tifffile==2025.5.10
tokenizer==3.4.5
tokenizers==0.19.1
torch==2.7.0+cu118
torch-fidelity==0.3.0
torchaudio==2.7.0+cu118
torchlibrosa==0.1.0
torchmetrics==1.7.3
torchvision==0.22.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
typing_inspect @ file:///home/conda/feedstock_root/build_artifacts/typing_inspect_1733845867367/work
tzdata==2025.2
undecorated==0.3.0
urllib3 @ file:///croot/urllib3_1737133630106/work
wadler_lindig @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_wadler-lindig_1744605034/work
wandb==0.19.11
wcwidth==0.2.13
wrapt @ file:///croot/wrapt_1736540904746/work
xxhash==3.5.0
yarl==1.19.0
zipp @ file:///croot/zipp_1732630741423/work
